Task fixes
